Feedback: - Renaming applicable -> isApplicable. - Removing unnecessary annotation. - Rename ValidationUtils -> AstValidationUtils. - Rename MuleSdkLocalExtensionModelsEnricher -> MuleSdkExtensionExtensionModelsEnricher
