+++ modules/artifact-activation/src/main/java/org/mule/runtime/module/artifact/activation/internal/ast/ApplicationExtensionModelsEnricher.java
@@ -58,7 +58,7 @@ public class ApplicationExtensionModelsEnricher implements ArtifactExtensionMode
   }
 
   @Override
-  public boolean applicable(ArtifactAst ast) {
+  public boolean isApplicable(ArtifactAst ast) {
     return ast.topLevelComponentsStream()
         .anyMatch(component -> REUSABLE_COMPONENT_TYPES.contains(component.getComponentType()));
   }
