+++ modules/artifact-activation/src/main/java/org/mule/runtime/module/artifact/activation/internal/ast/ArtifactAstUtils.java
@@ -6,7 +6,6 @@
  */
 package org.mule.runtime.module.artifact.activation.internal.ast;
 
-import org.mule.api.annotation.NoInstantiate;
 import org.mule.runtime.api.meta.model.ExtensionModel;
 import org.mule.runtime.api.metadata.ExpressionLanguageMetadataService;
 import org.mule.runtime.ast.api.ArtifactAst;
@@ -22,7 +21,6 @@ import java.util.Set;
  *
  * @since 4.5.0
  */
-@NoInstantiate
 public class ArtifactAstUtils {
 
   /**
@@ -52,7 +50,7 @@ public class ArtifactAstUtils {
 
     final ArtifactAst partialAst = doParseArtifactIntoAst(configResources, parserSupplier, extensions, true);
 
-    if (extensionModelsEnricher.applicable(partialAst)) {
+    if (extensionModelsEnricher.isApplicable(partialAst)) {
       Set<ExtensionModel> enrichedExtensionModels =
           extensionModelsEnricher.getEnrichedExtensionModels(partialAst, artifactClassLoader, extensions);
       if (!enrichedExtensionModels.equals(extensions)) {
